+++ sdk/resourcemanager/alertsmanagement/armalertsmanagement/alerts_client.go
@@ -18,7 +18,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/streaming"
 	"net/http"
 	"net/url"
 	"strconv"
@@ -99,8 +98,7 @@ func (client *AlertsClient) changeStateCreateRequest(ctx context.Context, alertI
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.Comment != nil {
-		body := streaming.NopCloser(strings.NewReader(*options.Comment))
-		return req, req.SetBody(body, "application/json")
+		return req, runtime.MarshalAsJSON(req, *options.Comment)
 	}
 	return req, nil
 }
@@ -117,7 +115,6 @@ func (client *AlertsClient) changeStateHandleResponse(resp *http.Response) (Aler
 // NewGetAllPager - List all existing alerts, where the results can be filtered on the basis of multiple parameters (e.g.
 // time range). The results can then be sorted on the basis specific fields, with the default being
 // lastModifiedDateTime.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2019-05-05-preview
 // options - AlertsClientGetAllOptions contains the optional parameters for the AlertsClient.GetAll method.
 func (client *AlertsClient) NewGetAllPager(options *AlertsClientGetAllOptions) *runtime.Pager[AlertsClientGetAllResponse] {